+++ src/ios/Sonogram.c
@@ -22,7 +22,7 @@ void Sonogram_initialise(Sonogram *sonogram, int numberOfFilters, int startFrequ
 
     for (int i = 0; i < numberOfFilters; i += 1) {
 
-        sonogram->goerztelFilters[i] = GoertzelFilter_initialise(GOERTZEL_WINDOW_SIZE, (float)startFrequency + (float) i * (float)frequencyStep, SAMPLES_PER_SECOND);
+        sonogram->goerztelFilters[i] = GoertzelFilter_initialise(GOERTZEL_WINDOW_SIZE, (double)startFrequency + (double) i * (double)frequencyStep, SAMPLES_PER_SECOND);
 
     }
 
@@ -69,7 +69,7 @@ void Sonogram_update(SInt16 sample, Sonogram *sonogram) {
 
 }
 
-float Sonogram_getCurrentValue(Sonogram *sonogram, int filter) {
+double Sonogram_getCurrentValue(Sonogram *sonogram, int filter) {
 
     return GoertzelFilter_estimate(&sonogram->goerztelFilters[filter]);
 
@@ -93,9 +93,9 @@ void Sonogram_copySonogram(Sonogram *sonogram) {
 
         for (int i = 0; i < sonogram->numberOfFilters; i += 1) {
 
-            memcpy(&sonogram->copySonogram[i][0], &sonogram->mainSonogram[i][firstSectionLength], secondSectionLength * sizeof(float));
+            memcpy(&sonogram->copySonogram[i][0], &sonogram->mainSonogram[i][firstSectionLength], secondSectionLength * sizeof(double));
 
-            memcpy(&sonogram->copySonogram[i][secondSectionLength], &sonogram->mainSonogram[i][0], firstSectionLength * sizeof(float));
+            memcpy(&sonogram->copySonogram[i][secondSectionLength], &sonogram->mainSonogram[i][0], firstSectionLength * sizeof(double));
 
         }
 
@@ -103,7 +103,7 @@ void Sonogram_copySonogram(Sonogram *sonogram) {
 
         for (int i = 0; i < sonogram->numberOfFilters; i += 1) {
 
-            memcpy(sonogram->copySonogram[i], sonogram->mainSonogram[i], SONOGRAM_LENGTH * sizeof(float));
+            memcpy(sonogram->copySonogram[i], sonogram->mainSonogram[i], SONOGRAM_LENGTH * sizeof(double));
 
         }
 
@@ -111,7 +111,7 @@ void Sonogram_copySonogram(Sonogram *sonogram) {
 
 }
 
-float Sonogram_getValue(Sonogram *sonogram, int i, int j, int width, int height, int duration) {
+double Sonogram_getValue(Sonogram *sonogram, int i, int j, int width, int height, int duration) {
 
     if (i < 0 || i > width-1 || j < 0 || j > height-1) {
 
@@ -121,13 +121,13 @@ float Sonogram_getValue(Sonogram *sonogram, int i, int j, int width, int height,
 
         int numberOfMeasurementPoints = duration * SAMPLES_PER_SECOND / GOERTZEL_WINDOW_SIZE;
 
-        float xRatio = (float)(numberOfMeasurementPoints - 1) / (float)(width - 1);
+        double xRatio = (float)(numberOfMeasurementPoints - 1) / (double)(width - 1);
 
-        float yRatio = (float)(sonogram->numberOfFilters - 1) / (float)(height - 1);
+        double yRatio = (float)(sonogram->numberOfFilters - 1) / (double)(height - 1);
 
-        int xIndex = SONOGRAM_LENGTH - numberOfMeasurementPoints + (int)(0.5f + (float)i * xRatio);
+        int xIndex = SONOGRAM_LENGTH - numberOfMeasurementPoints + (int)(0.5 + (double)i * xRatio);
 
-        int yIndex = sonogram->numberOfFilters - 1 - (int)(0.5f + (float)j * yRatio);
+        int yIndex = sonogram->numberOfFilters - 1 - (int)(0.5 + (double)j * yRatio);
 
         return sonogram->copySonogram[yIndex][xIndex];
 
@@ -135,13 +135,13 @@ float Sonogram_getValue(Sonogram *sonogram, int i, int j, int width, int height,
 
 }
 
-float interpolate(float value, float x0, float x1, float y0, float y1) {
+float interpolate(double value, double x0, double x1, double y0, double y1) {
 
     return (value - x0) * (y1 - y0) / (x1 - x0) + y0;
 
 }
 
-float base(float value) {
+double base(double value) {
 
     if (value <= -0.75) {
         return 0;
@@ -157,14 +157,14 @@ float base(float value) {
 
 }
 
-void Sonogram_colour(float value, float min, float max, UInt8 *red, UInt8 *green, UInt8 *blue) {
+void Sonogram_colour(double value, double min, double max, UInt8 *red, UInt8 *green, UInt8 *blue) {
 
-    float scaledValue = MIN(1.0f, MAX(0.0f, logf( (value + 1.0f) / min) / logf(max / min)));
+    double scaledValue = MIN(1.0, MAX(0.0, logf( (value + 1.0) / min) / logf(max / min)));
 
-    *red = (UInt8) (0.5f + 255.0f * base(scaledValue - 0.5f));
+    *red = (UInt8) (0.5 + 255.0 * base(scaledValue - 0.5));
 
-    *green = (UInt8) (0.5f + 255.0f * base(scaledValue));
+    *green = (UInt8) (0.5 + 255.0 * base(scaledValue));
 
-    *blue = (UInt8) (0.5f + 255.0f * base(scaledValue + 0.5f));
+    *blue = (UInt8) (0.5 + 255.0 * base(scaledValue + 0.5));
 
 }
